# docs/api/swagger.yaml
components:
  schemas:
    AgentConfig:
      properties:
        approvalTimeout:
          example: 24h
          type: string
        maxConcurrency:
          example: 3
          type: integer
        taskRetryLimit:
          example: 3
          type: integer
        type:
          example: claude-code
          type: string
      type: object
    AgentConfigUpdate:
      properties:
        approvalTimeout:
          type: string
        maxConcurrency:
          type: integer
        taskRetryLimit:
          type: integer
      type: object
    ApiError:
      properties:
        code:
          example: BAD_REQUEST
          type: string
        details:
          description: |-
            Details contains additional error context (field names, validation rules, etc.)
            Type varies by error code: map[string]string for validation, string for general errors
          example: '{"field":"email","rule":"required"}'
        message:
          example: Invalid request parameters
          type: string
      type: object
    AuditConfigDTO:
      properties:
        enabled:
          example: true
          type: boolean
        maxOutputLength:
          example: 1000
          type: integer
        retentionDays:
          example: 90
          type: integer
      type: object
    AuditConfigUpdate:
      properties:
        enabled:
          type: boolean
        maxOutputLength:
          type: integer
        retentionDays:
          type: integer
      type: object
...
    ClarityConfig:
      properties:
        autoProceedThreshold:
          example: 80
          type: integer
        forceClarifyThreshold:
          example: 40
          type: integer
        threshold:
          example: 60
          type: integer
      type: object
    ClarityConfigUpdate:
      properties:
        autoProceedThreshold:
          type: integer
        forceClarifyThreshold:
          type: integer
        threshold:
          type: integer
      type: object
    ComplexityConfig:
      properties:
        forceDesignConfirm:
          example: false
          type: boolean
        threshold:
          example: 70
          type: integer
      type: object
    ComplexityConfigUpdate:
      properties:
        forceDesignConfirm:
          type: boolean
        threshold:
          type: integer
      type: object
    Config:
      properties:
        agent:
          $ref: '#/components/schemas/AgentConfig'
        audit:
          $ref: '#/components/schemas/AuditConfigDTO'
        clarity:
          $ref: '#/components/schemas/ClarityConfig'
        complexity:
          $ref: '#/components/schemas/ComplexityConfig'
        database:
          $ref: '#/components/schemas/DatabaseConfig'
        git:
          $ref: '#/components/schemas/GitConfig'
        server:
          $ref: '#/components/schemas/ServerConfig'
        webhook:
          $ref: '#/components/schemas/WebhookConfig'
      type: object
...
    DatabaseConfig:
      properties:
        connMaxLifetime:
          example: 1h
          type: string
        host:
          example: localhost
          type: string
        maxIdleConns:
          example: 10
          type: integer
        maxOpenConns:
          example: 100
          type: integer
        name:
          example: litchi
          type: string
        port:
          example: 5432
          type: integer
        sslmode:
          example: disable
          type: string
        user:
          example: postgres
          type: string
      type: object
...
    GitConfig:
      properties:
        branchNamingPattern:
          example: issue-{number}-{slug}
          type: string
        commandTimeout:
          example: 5m
          type: string
        commitSignOff:
          example: true
          type: boolean
        defaultBaseBranch:
          example: main
          type: string
        gitBinaryPath:
          example: git
          type: string
        worktreeAutoClean:
          example: true
          type: boolean
        worktreeBasePath:
          example: /var/litchi/worktrees
          type: string
      type: object
    GitConfigUpdate:
      properties:
        branchNamingPattern:
          type: string
        commandTimeout:
          type: string
        commitSignOff:
          type: boolean
        defaultBaseBranch:
          type: string
        worktreeAutoClean:
          type: boolean
        worktreeBasePath:
          type: string
      type: object
...
    ServerConfig:
      properties:
        host:
          example: 0.0.0.0
          type: string
        mode:
          example: debug
          type: string
        port:
          example: 8080
          type: integer
        version:
          example: 0.1.0
          type: string
      type: object
...
    UpdateConfig:
      properties:
        agent:
          $ref: '#/components/schemas/AgentConfigUpdate'
        audit:
          $ref: '#/components/schemas/AuditConfigUpdate'
        clarity:
          $ref: '#/components/schemas/ClarityConfigUpdate'
        complexity:
          $ref: '#/components/schemas/ComplexityConfigUpdate'
        git:
          $ref: '#/components/schemas/GitConfigUpdate'
      type: object
...
    WebhookConfig:
      properties:
        idempotencyAutoCleanup:
          example: true
          type: boolean
        idempotencyEnabled:
          example: true
          type: boolean
        idempotencyTTL:
          example: 24h
          type: string
      type: object
...
info:
  description: Automated development agent system - from GitHub Issue to Pull Request
  title: Litchi API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/config:
    get:
      description: Returns the current application configuration (sensitive fields
        excluded)
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Config'
          description: Configuration retrieved successfully
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Internal server error
      summary: Get configuration
      tags:
      - config
    put:
      description: Updates application configuration (only certain fields are updatable
        at runtime)
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/UpdateConfig'
                description: Configuration update request
                summary: request
        description: Configuration update request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Config'
          description: Configuration updated successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Invalid request or validation failed
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Internal server error
      summary: Update configuration
      tags:
      - config
...